Ziglang branch of trOS
This is a new branch that features a rewrite of trOS in Zig. The current
feature parity is 1:1, everything that works in the Rust branch works in
Zig. I will be deciding soon which one I will work on full time, which
means the other branch that isn't chosen will be relegated to
backporting every once and a while.